requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.31.0
alpaca-py>=0.30.0
python-dotenv>=0.13.0